% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstrasAlgorithm.R
\docType{package}
\name{dijkstrasAlgorithm-package}
\alias{dijkstrasAlgorithm}
\alias{dijkstrasAlgorithm-package}
\title{Euclidean and Dijkstra's algorithm as an R package.}
\description{
An R package for solving the Euclidean and Dijkstra algorithms. It includes a function for computing the greatest common divisor between two scalars (Euclidean algorithm) and a function for computing the shortest paths between nodes in a graph (Dijkstra's algorithm).
}
\author{
\strong{Maintainer}: Theodor Emanuelsson \email{theodor.emanuelsson@me.com}

}
