% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{The Euclidean algorithm: Greatest common divisor of two scalars}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A natural number.}

\item{b}{A natural number.}
}
\value{
Returns the greatest common divisor of a and b
}
\description{
The Euclidean algorithm: Greatest common divisor of two scalars
}
\examples{
euclidean(123612, 13892347912) # = 4
euclidean(100, 1000) # = 100
}
